Move tests for util functions
